# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAAS API",
        "description": "API documentation for SAAS Dashboard",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/create_course": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "إنشاء كورس جديد",
                "operationId": "createCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "stage_and_subject_id",
                                    "price",
                                    "is_free"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mathematics 101"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Basic Math course for beginners"
                                    },
                                    "stage_and_subject_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "is_free": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم إنشاء الكورس بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم إنشاء الكورس بنجاح"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/update_course": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "تحديث بيانات كورس",
                "operationId": "updateCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Course Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "stage_and_subject_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 120.5
                                    },
                                    "is_free": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تحديث الكورس بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم تحديث الكورس بنجاح"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fetch_course_details": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "جلب تفاصيل كورس محدد",
                "operationId": "fetchCourseDetails",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم جلب الكورس بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم جلب الكورس بنجاح"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/delete_course": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "حذف كورس",
                "operationId": "deleteCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم حذف الكورس بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم حذف الكورس بنجاح"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/is_active": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "تغيير حالة التفعيل لكورس",
                "operationId": "toggleCourseActive",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تغيير حالة الكورس بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم تغيير حالة الكورس بنجاح"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/order_course": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "تغيير ترتيب الكورس",
                "operationId": "orderCourse",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_id",
                                    "order"
                                ],
                                "properties": {
                                    "course_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "تم تغيير ترتيب الكورس بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم تغيير ترتيب الكورس بنجاح"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CourseResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/fetch_all_courses": {
            "post": {
                "tags": [
                    "Courses"
                ],
                "summary": "جلب جميع الكورسات",
                "operationId": "fetchAllCourses",
                "parameters": [
                    {
                        "name": "word",
                        "in": "query",
                        "description": "للبحث باسم الكورس",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "with_paginate",
                        "in": "query",
                        "description": "تفعيل أو إيقاف pagination (1 أو 0)",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "عدد النتائج لكل صفحة في حالة pagination",
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم جلب جميع الكورسات بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "تم جلب جميع الكورسات بنجاح"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CourseResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CourseResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Mathematics 101"
                    },
                    "description": {
                        "type": "string",
                        "example": "Basic Math course"
                    },
                    "order": {
                        "type": "integer",
                        "example": 1
                    },
                    "stage_and_subject_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 100.5
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_free": {
                        "type": "boolean",
                        "example": false
                    },
                    "organization_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use bearer token to access this API",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}